Change semantics of withAsync & double performance of race and concurrently
Previously in "withAsync action inner" when "inner" would receive an
asynchronous exception "action" would be killed. Now "action" will
receive the same asynchronous exception has "inner". It's still the case
that when"inner" receives a synchronous exception "action" will be
killed.

This change in semantics allows for an implementation of "race" and
"concurrently" which is twice as fast because we can now fork a single
thread instead of two.

This patch was co-written with my colleague @asayers.