docs: Explain that `withAsync` should be preferred over `async`.
Over the last 5 years I have seen way too many people use `async`
incorrectly because they do not read or remember the tutorial,
and the temptation is large to copy-paste the first code example
that appears in the documentation (some might say that is
unconvientional/surprising to have the first code example be
not-production-ready).

Also, it is unintuitive to beginners that the function `async`
(after which the package is even named) is the function they should
_avoid_ in practice, so it deserves an extra comment.

This commit makes the documentation more explicit, and also does away
with the "slight improvement" terminology on the code that is absolutely
necessary to write non-leaking production code.