docs: Introduce the safe functions before the details.
This implements Simon Marlow's feedback in PR #104:

> The documentation should start by introducing `withAsync`,
> and only later should talk about `async` and `wait`.
> It's a bigger restructuring, but I think starting the docs
> with some code that you shouldn't use is not the right thing at all

To make clear what functions people should usually use,
split the introduction into a "High-level" and "Low-level"
API, which is terminology I found very successful when it comes
to explaining when which function should be used.

Also add detail and improve markup in a few places.